fix: fp32 IO for gte exports; docs for intel OVMS fetch/export + GPU-proof reload procedure
- export script: load torch models with dtype=float32 (thenlper/gte-* ship
  fp16 weights; tracing those made OVMS answer Embed with FP16, which the
  ovms backend rejects). Re-exported + re-pinned gte-small/gte-base.
- docs/adding-ovms-embedding-pipelines.md: export via make
  fetch-embeddings-intel, Parameter_1 tokenizer input pinning, fp32 note,
  ONNX-surgery path for nomic, no-restart reload procedure incl. the
  in-place-file-change gotcha, and GPU placement verification (target
  device log lines + xe drm-cycles-ccs fdinfo sampling).
- docs/fetching-models.md: intel row now fully pinned + hashed.

+++ scripts/export_ovms_embeddings.py
@@ -187,7 +187,12 @@ def export_alias(alias: str, revision: str, out: Path, hf_out: Path) -> None:
         example = hf_tok(["inferstream export smoke"], return_tensors="pt")
         return finish_export(alias, spec, ov_model, hf_tok, example, out, hf_out)
 
-    model = AutoModel.from_pretrained(repo, revision=revision)
+    # dtype=float32 explicitly: transformers 5.x defaults to the checkpoint
+    # dtype, and some repos (e.g. thenlper/gte-*) ship fp16 weights — tracing
+    # those yields an fp16-in/out graph, and OVMS then answers Embed with an
+    # FP16 tensor the inferstream backend rejects. FP16 compression of the
+    # IR happens at save_model time regardless; graph IO must stay FP32.
+    model = AutoModel.from_pretrained(repo, revision=revision, dtype=torch.float32)
     model = model.eval()
 
     class Wrapper(torch.nn.Module):
